If the _updated_at field doesn't exist it will not save to a file the latest entry processed

+++ pkg/adaptor/elasticsearch/elasticsearch.go
@@ -135,17 +135,19 @@ func (e *Elasticsearch) applyOp(msg *message.Msg) (*message.Msg, error) {
 	}
 	delete(msg.Map(), "_id")
 	msg_map := msg.Map()
-	last_update_date := msg_map["_updated_at"]
-	date_string := fmt.Sprint(last_update_date)
-	f, err := os.Create("last_date")
-	if err != nil {
-		fmt.Println(err)
+	last_update_date, exists := msg_map["_updated_at"]
+	if exists {
+		date_string := fmt.Sprint(last_update_date)
+		f, err := os.Create("last_date")
+		if err != nil {
+			fmt.Println(err)
+		}
+		n, err := io.WriteString(f, date_string)
+		if err != nil {
+			fmt.Println(n, err)
+		}
+		f.Close()
 	}
-	n, err := io.WriteString(f, date_string)
-	if err != nil {
-		fmt.Println(n, err)
-	}
-	f.Close()
 	
 	_, _type, err := msg.SplitNamespace()
 	if err != nil {
